Fix .travis errors

+++ mx.irbuilder/mx_irbuilder.py
@@ -207,9 +207,10 @@ def runIRBuilderTestGen38(vmArgs):
         """runs the test suite"""
 
         # remove old files inside build directory
-        for file in os.listdir(suite[1]):
-            if file.endswith(".ll") or file.endswith(".bc"):
-                os.remove(os.path.join(suite[1], file))
+        if os.path.isdir(suite[1]):
+            for file in os.listdir(suite[1]):
+                if file.endswith(".ll") or file.endswith(".bc"):
+                    os.remove(os.path.join(suite[1], file))
 
         try:
             mx_sulong.mx_testsuites.run(vmArgs + ['-Dirwriter.LLVMVersion=3.8'], suite[0], [])
